(PC-27674)[PRO] feat: move dms timeline to collective tab
