--- pro/src/components/VenueForm/CollectiveVenueInformations/__specs__/CollectiveVenueInformationsEdition.spec.tsx
@@ -1,81 +1,0 @@
-import { screen } from '@testing-library/react'
-import { addDays } from 'date-fns'
-import React from 'react'
-
-import { DMSApplicationstatus } from 'apiClient/v1'
-import {
-  defaultCollectiveDmsApplication,
-  defaultVenue,
-} from 'utils/collectiveApiFactories'
-import { renderWithProviders } from 'utils/renderWithProviders'
-
-import {
-  CollectiveVenueInformationsEdition,
-  CollectiveVenueInformationsEditionProps,
-} from '../CollectiveVenueInformationsEdition'
-
-const renderCollectiveVenueInformations = (
-  props: CollectiveVenueInformationsEditionProps
-) => {
-  renderWithProviders(<CollectiveVenueInformationsEdition {...props} />)
-}
-
-describe('CollectiveVenueInformations', () => {
-  it('should display description when venue is not accepted on adage', () => {
-    renderCollectiveVenueInformations({
-      venue: {
-        ...defaultVenue,
-        collectiveDmsApplication: {
-          ...defaultCollectiveDmsApplication,
-          state: DMSApplicationstatus.EN_CONSTRUCTION,
-        },
-      },
-      canCreateCollectiveOffer: false,
-    })
-    expect(
-      screen.getByText(
-        'Pour publier des offres à destination des scolaires, votre lieu doit être référencé sur ADAGE, la plateforme dédiée aux enseignants et aux chefs d’établissements.'
-      )
-    ).toBeInTheDocument()
-  })
-
-  it('should display timeline if venue has adage id since less than 30 days', () => {
-    renderCollectiveVenueInformations({
-      venue: {
-        ...defaultVenue,
-        hasAdageId: true,
-        adageInscriptionDate: addDays(new Date(), -10).toISOString(),
-        collectiveDmsApplication: {
-          ...defaultCollectiveDmsApplication,
-          state: DMSApplicationstatus.ACCEPTE,
-        },
-      },
-      canCreateCollectiveOffer: true,
-    })
-    expect(
-      screen.getByText(
-        'Votre lieu a été réferencé dans ADAGE par les équipes du Ministère de l’Education Nationale'
-      )
-    ).toBeInTheDocument()
-  })
-
-  it('should display eac section if venue has adage id since more than 30 days', () => {
-    renderCollectiveVenueInformations({
-      venue: {
-        ...defaultVenue,
-        hasAdageId: true,
-        adageInscriptionDate: addDays(new Date(), -32).toISOString(),
-        collectiveDmsApplication: {
-          ...defaultCollectiveDmsApplication,
-          state: DMSApplicationstatus.ACCEPTE,
-        },
-      },
-      canCreateCollectiveOffer: true,
-    })
-    expect(
-      screen.getByRole('heading', {
-        name: 'Mes informations pour les enseignants',
-      })
-    ).toBeInTheDocument()
-  })
-})
